refactor: move Serializer to separate namespace
